+++ src/patient_representation/tl/__init__.py
@@ -3,6 +3,7 @@ from .basic import (
     PILOT,
     CellGroupComposition,
     DiffusionEarthMoverDistance,
+    GloScope,
     GroupedPseudobulk,
     MrVI,
     Pseudobulk,